feat(database): migrate to BigInt for IDs and enhance schema
refactor(models): update ID columns to BigInt for scalability
fix(migrations): remove deprecated migration files
perf(alembic): optimize migration execution with timeouts
docs(models): improve model documentation and imports
style: clean up removed files and update requirements

+++ app/infrastructure/database/models/v2/base.py
@@ -2,7 +2,7 @@
 Base SQLAlchemy model for V2 database schema
 """
 from datetime import datetime
-from sqlalchemy import Column, Integer, DateTime
+from sqlalchemy import Column, BigInteger, DateTime
 from sqlalchemy.orm import declarative_base
 from sqlalchemy.sql import func
 
@@ -14,7 +14,7 @@ class BaseModel(Base):
     """
     __abstract__ = True
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now(), nullable=False)
