+++ app/infrastructure/database/models/v2/user_skill_progress.py
@@ -5,7 +5,7 @@ This model serves as a join table to track which SkillTreeNodes have been
 unlocked by each Ascendant. It represents the progression state of users
 through the skill tree system.
 """
-from sqlalchemy import Column, Integer, ForeignKey, DateTime, Index, UniqueConstraint
+from sqlalchemy import Column, Integer, BigInteger, ForeignKey, DateTime, Index, UniqueConstraint
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -26,7 +26,7 @@ class UserSkillProgress(BaseModel):
     __tablename__ = 'user_skill_progress'
     
     # Foreign key relationships
-    ascendant_id = Column(Integer, ForeignKey('ascendants.id'), nullable=False)
+    ascendant_id = Column(BigInteger, ForeignKey('ascendants.id'), nullable=False)
     node_id = Column(Integer, ForeignKey('skill_tree_nodes.id'), nullable=False)
     
     # Tracking information
